docs: move files
