+++ src/app/layout.tsx
@@ -3,10 +3,10 @@ import './globals.css';
 import { AntdRegistry } from '@ant-design/nextjs-registry';
 import { ConfigProvider } from 'antd';
 
+import Header from '@/components/header/header';
 import { customTheme } from '@/lib/antd/theme';
 
 import StoreProvider from './StoreProvider';
-import Header from './_components/header/header';
 
 import type { Metadata } from 'next';
